requirements.txt
# --- Web stack (tương thích) ---
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7

# --- Core libs của dự án---
googletrans==4.0.0rc1
numpy==1.23.3
pandas==1.5.0
opencv-python==4.6.0.66
underthesea==1.3.5a3
pyvi==0.1.1
translate==3.6.1
faiss-cpu==1.7.2
ftfy==6.1.1
regex==2022.9.13
tqdm==4.64.1
matplotlib==3.6.0
langdetect==1.0.9
rapidfuzz==2.12.0
sentence_transformers==2.2.2
Pillow>=9.2.0

# --- Xuất Excel ---
openpyxl>=3.1

# --- CLIP (yêu cầu PyTorch đã cài thủ công phù hợp hệ thống) ---
git+https://github.com/openai/CLIP.git

# --- TRAKE (pose & DTW) ---
# mediapipe: pose extraction backend (optional but recommended for TRAKE)
mediapipe==0.10.11
# fastdtw: fast Dynamic Time Warping implementation
fastdtw==0.3.4
# scipy: distance metrics for DTW and general utilities
scipy==1.10.1
